requirements.txt
python-dotenv==1.0.0
PyPDF2==3.0.1
requests==2.31.0
numpy==1.24.3
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
pgvector==0.2.4 